chore: Rename app name to dapurNgebul
