+++ android/app/BUCK
@@ -45,12 +45,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.akveo.kittenTricks",
+    package = "com.akveo.dapurNgebul",
 )
 
 android_resource(
     name = "res",
-    package = "com.akveo.kittenTricks",
+    package = "com.akveo.dapurNgebul",
     res = "src/main/res",
 )
