+++ android/app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    package="com.akveo.kittenTricks"
+    package="com.akveo.dapurNgebul"
     android:versionCode="15"
     android:versionName="1.0.9">
