+++ android/app/src/main/java/com/akveo/dapurNgebul/MainApplication.java
@@ -1,4 +1,4 @@
-package com.akveo.kittenTricks;
+package com.akveo.dapurNgebul;
 
 import android.app.Application;
