+++ android/app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
     <string moduleConfig="true" name="mobileCenterCrashes_whenToSendCrashes">ALWAYS_SEND</string>
     <string moduleConfig="true" name="mobileCenterAnalytics_whenToEnableAnalytics">ALWAYS_SEND</string>
-    <string name="app_name">Kitten Tricks</string>
+    <string name="app_name">Dapur Ngebul</string>
 </resources>
