+++ android/gradle.properties
@@ -18,5 +18,5 @@
 # org.gradle.parallel=true
 
 android.useDeprecatedNdk=true
-RELEASE_STORE_FILE=kittenTricks.jks
-RELEASE_KEY_ALIAS=kittenTricks
+RELEASE_STORE_FILE=dapurNgebul.jks
+RELEASE_KEY_ALIAS=dapurNgebul
