+++ android/settings.gradle
@@ -1,4 +1,4 @@
-rootProject.name = 'akveo.kittenTricks'
+rootProject.name = 'akveo.dapurNgebul'
 include ':realm'
 project(':realm').projectDir = new File(rootProject.projectDir, '../node_modules/realm/android')
 include ':react-native-google-analytics-bridge'
