+++ app.json
@@ -1,4 +1,4 @@
 {
-  "name": "kittentricks",
-  "displayName": "Kitten Tricks"
+  "name": "dapurNgebul",
+  "displayName": "Dapur Ngebul"
 }
