+++ ios/dapurNgebul.xcworkspace/contents.xcworkspacedata
@@ -2,6 +2,6 @@
 <Workspace
    version = "1.0">
    <FileRef
-      location = "self:/Users/alexei/dev/kittenTricks/ios/kittenTricks.xcodeproj">
+      location = "group:dapurNgebul.xcodeproj">
    </FileRef>
 </Workspace>
