+++ ios/dapurNgebul/Base.lproj/LaunchScreen.xib
