+++ ios/dapurNgebul/Images.xcassets/AppIcon.appiconset/icon-20@1x.png
