+++ ios/dapurNgebul/Images.xcassets/AppIcon.appiconset/icon-50@1x.png
