+++ ios/dapurNgebul/Images.xcassets/AppIcon.appiconset/icon-50@2x.png
