+++ ios/dapurNgebul/Images.xcassets/AppIcon.appiconset/icon-57@1x.png
