+++ ios/dapurNgebul/Images.xcassets/AppIcon.appiconset/icon-57@2x.png
