+++ ios/dapurNgebul/Images.xcassets/AppIcon.appiconset/icon-60@2x.png
