+++ ios/dapurNgebul/Images.xcassets/AppIcon.appiconset/icon-72@1x.png
