+++ ios/dapurNgebul/Images.xcassets/AppIcon.appiconset/icon-76@1x.png
