+++ ios/dapurNgebul/Images.xcassets/AppIcon.appiconset/icon-76@2x.png
