+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "kittenTricks",
+  "name": "dapurNgebul",
   "version": "1.0.9",
   "private": true,
   "author": "akveo",
